Rip out dotenv dependency which was only used in dev.
dotenv is not really the right solution to our configuration needs since
Aptible has it's own system for sharing env variables with the building
and running containers. The project already only used it for
development, so no reason the have it as a dependency when people can
just use their own local solution for managing env variables per
directory. I highly recommend direnv ;)