Cron message scheduler can recover if no message is found for patient.
Previously any inability to find a message for a particular patient
would crash the scheduler and not send messages to any patients.

TBH commit is not nearly enough to make the message scheduler durable to
failure, but this is a quick hacky fix to get staging working again for
product validation.