Auto fix prettier issues
These issues were not being flagged by eslint previously, but now they
are.

There is 1 issue that is some janky typing that is in turn causing
prettier issues. I've gotten around this by having eslint just warn on
prettier issues, but eventually we should fix that and then turn on
prettier issues as eslint errors